Update README.md
This repository contains test scripts written in Selenium Java POM. In order to run these scripts you need to open the src/test/java folder and when you open any .java file yo will need to run the tests by clicking on the Run below the @Test of a Method. Some times if the Run is not visible just right click on the mouse, select Run as, Run configurations, choose the Class and the Method   and click Run. After that the script will be executed.